requirements.txt
openai
python-dotenv
pydantic
llama-index-core
llama-index-readers-file
llama-index-vector-stores-faiss
llama-index-vector-stores-qdrant
llamaindex-py-client
qdrant-client
requests
faiss-cpu

chromadb
llama_index


langchain
langchain_community